% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FathomR_Functions.R, R/auth.R
\name{fathom_authenticate}
\alias{fathom_authenticate}
\title{Authenticate to Fathom API}
\usage{
fathom_authenticate(username, password, client_id)

fathom_authenticate(username, password, client_id)
}
\arguments{
\item{username}{Fathom account email}

\item{password}{Fathom password}

\item{client_id}{AWS Cognito client ID}
}
\value{
A string token used to authenticate other requests

Access token string
}
\description{
This function logs into the Fathom API using the provided credentials and returns an access token.

Logs into the Fathom Cognito authentication system and returns an access token.
}
