% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fathom_to_vue.R
\name{fathom_to_vue}
\alias{fathom_to_vue}
\title{Fathom to VUE}
\usage{
fathom_to_vue(det, cfg_station, spec_sheet = NULL)
}
\arguments{
\item{det}{tibble containing detection data generated by parse_fathom_files()}

\item{cfg_station}{tibble containing station configuration data generated by parse_fathom_files()}

\item{spec_sheet}{tibble containing transmitter specification generated using glatos::read_vemco_tag_specs()}
}
\value{
Returns a tibble containing data formatted to match the strucuture of detection data exported from VUE software
Tibble also contains a 'Time.corrected' variable which in the Date.and.Time..UTC. following lineaer time correction to
compensate for receiver clock drift
}
\description{
Helper function to convert detection data parsed by parse_fathom_files() to replicate the data structure of detections exported from legacy Vemco User Environment (VUE) software.
}
\examples{

\dontrun{
#'   # This example shows how to call the process_log_files function
#'   # Replace "/path/to/your/data" with a valid directory path on your system
#'   parse_fathom_files('/path_to_fathom_csv_files', rec_type = c("DET, "CFG_STATION"), output = 'env")

#'   # without  transmitter spec_sheet
#'   x <- fathom_to_vue_conversion(det, cfg_station)

#'   # with transmitter spec_sheet

#' spec_files <- list.files('path/to/spec_sheet', pattern='.xls', full.names = TRUE)
#'     spec_sheet <- lapply(spec_files, read_vemco_tag_specs) \%>\%
#'     map_dfr(~as_tibble(.x$specs)) \%>\%
#'     mutate(Transmitter = paste0(code_space,'-',id_code))
#'   x <- fathom_to_vue_conversion(det, cfg_station, spec_sheet = spec_sheet)

#' }


}
