% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detections.R
\name{get_detections}
\alias{get_detections}
\title{Get detections from Fathom Central}
\usage{
get_detections(
  common_names = "all",
  transmitters = "all",
  transmitterTypes = "all",
  study = "all",
  start_date = NULL,
  end_date = NULL,
  token = NULL,
  ws_id = NULL
)
}
\arguments{
\item{common_names}{Character vector of species names or "all"}

\item{transmitters}{Character vector of transmitter IDs or "all"}

\item{transmitterTypes}{Character vector of tag types or "all"}

\item{study}{Study name(s) or "all"}

\item{start_date}{Start date (UTC) or NULL}

\item{end_date}{End date (UTC) or NULL}

\item{token}{API token (optional)}

\item{ws_id}{workspace ID (optional)}
}
\value{
A tibble of detections
}
\description{
Retrieves acoustic telemetry detection data from Fathom Central with optional filtering.
}
\examples{
\dontrun{
auth <- authenticate_wrapper()
detections_all <- get_detections(
#common_names = c("Cownose","Gulf Sturgeon"), # specify common name
transmitters = c("A69-1604-60495","A69-1303-46570","A69-9001-16605"), # For example: To Specify specific Transmitters
#transmitterTypes = c("V13","V16"),  # For example: to only get V13 and V16 transmitter types
#study = "CFRD", # name of study
#start_date = "2025-01-01", # use "all" or leave unspecified for all dates, or specify a start date in yyyy-mm-dd
#end_date = "2025-06-01", # use "all" or leave unspecified for all dates, or specify an end date in yyyy-mm-dd
token = auth$token,  # if not specified then authentication will prompt a new login
ws_id = auth$ws_id  # if not specified then authentication will prompt a new login
)
}
}
