% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biometrics.R
\name{get_event_biometrics}
\alias{get_event_biometrics}
\title{Get event-level biometrics dataset}
\usage{
get_event_biometrics(token = NULL, ws_id = NULL, species = "all")
}
\arguments{
\item{token}{Authentication token}

\item{ws_id}{Workspace ID}

\item{species}{Optional species filter ("all" = no filter)}
}
\description{
Returns one row per event with associated measurements and transmitter info.
}
