% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{get_workspaces}
\alias{get_workspaces}
\title{Get available workspaces}
\usage{
get_workspaces(token)
}
\arguments{
\item{token}{Authentication token}
}
\value{
Tibble of workspace IDs and names
}
\description{
Queries Fathom GraphQL API and returns accessible workspaces.
}
