% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_imos_meta.R
\name{match_imos_meta}
\alias{match_imos_meta}
\title{Match IMOS receiver metadata to parsed Fathom files}
\usage{
match_imos_meta(input, imos_meta)
}
\arguments{
\item{input}{List of files opened with parse_fathom_files()}

\item{imos_meta}{Path to the IMOS receiver deployment metadata csv file}
}
\value{
List of Fathom files, including IMOS metadata information. Please note that rows with no matching IMOS metadata will not be returned in the
matched outputs.
}
\description{
This function uses IMOS receiver metadata to add IMOS metadata, including receiver_deployment_id, installation_name, station_name, and deployment locations (receiver_deployment_latitude and receiver_deployment_longitude)
to the parsed Fathom files.
}
