% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fathom_files_vroom.R
\name{parse_fathom_files_vroom}
\alias{parse_fathom_files_vroom}
\title{Parse Fathom Files (vroom)}
\usage{
parse_fathom_files_vroom(directory_path)
}
\arguments{
\item{path}{path to the folder on your local machine containing the Fathom
exported log files (.csv)}
}
\description{
The function parses Innovasea's Fathom csv schema log files and returns a
dataframe for each record type.
}
\examples{

\dontrun{
#'   # This example shows how to call the process_log_files function
#'   # Replace "/path/to/your/data" with a valid directory path on your system
#'   process_log_files("/path/to/your/data")
#' }


}
