% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fathom_split_csv_files.R
\name{parse_fathom_split_csv_files}
\alias{parse_fathom_split_csv_files}
\title{Parse Fathom split csv Files}
\usage{
parse_fathom_split_csv_files(parent_directory, record_type = NULL)
}
\arguments{
\item{record_type}{string describing the record type to be parsed}

\item{path}{path to the parent folder containing subfolders for parsing}
}
\value{
Returns a dataframe for the selected record type
}
\description{
The function parses Innovasea's Fathom csv schema log files which have been
split into individual csv files based on record type
Fathom receiver log files (.vrl, .vdat) can be
converted to user readable formats using FathomR::convert_log_files() or
using Innovasea's Fathom Connect desktop application or cloud based service
Fathom Central. Note: Export formats from Fathom Central are restricted to
interleaved .csv.
}
\details{
The *Date Time (UTC)* time variable is referenced to UTC time zone and is
not adjusted for clock drift. *Time* variable is equivalent to the *Date
Time (UTC)* following correction for clock drift using the *Time Correction
(s)* variable.
}
\examples{

\dontrun{
#'   # This example shows how to call the process_log_files function
#'   # Replace "/path/to/your/data" with a valid directory path on your system
#'   parse_fathom_split_csv_files('/path_to_fathom_split_csv_files_parent_folder', record_type = 'DET')
#' }

}
