% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek_fathom_schema.R
\name{peek_fathom_schema}
\alias{peek_fathom_schema}
\title{Peek Fathom Schema}
\usage{
peek_fathom_schema(
  directory_path,
  output = c("list", "env"),
  envir = .GlobalEnv,
  overwrite = TRUE,
  n_max_header_lines = 500,
  example_n = 5,
  example_max_files = 50,
  verbose = TRUE
)
}
\arguments{
\item{directory_path}{path to the folder on your local machine containing the Fathom
interleaved .csv files}

\item{output}{defines the output type "list" or "env"}

\item{envir}{defines the environment to which dataframes will be added}

\item{overwrite}{will overwrite existing objects in the environment with the same name}

\item{n_max_header_lines}{integer defining the number of header lines to expect}

\item{example_n}{integer defining the number of rows to sample from each record type to populate into $report}

\item{example_max_files}{integer defining the max number of files to peek into}
}
\value{
Returns a list of containing two elements: $data and $report
$data contains blank example dataframes for each record type contained within the
interleaved Fathom csv files in the file path.
$report provides a list containing elements summarising the variables within each record type including example data defined by example_n
}
\description{
The function peeks into Innovasea's Fathom csv schema and returns a named list
containing example dataframes expected from parse_fathom_files().
}
\examples{

\dontrun{
#'   # This example shows how to call the process_log_files function
#'   # Replace "/path/to/your/data" with a valid directory path on your system
#'   fathom_schema <- peek_fathom_schema('/path_to_fathom_csv_files', output = 'list")

#'   fathom_schema <- peek_fathom_schema('/path_to_fathom_csv_files',
#'                                             output = 'list", example_n = 5)

#'  names(res$data)
#'  str(res$data$det)
#'  res$data$det
#'  print(n=30, res$report$summary)

#' }


}
