% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receiver_performance_report.R
\name{receiver_performance_report}
\alias{receiver_performance_report}
\title{Generate Automated Report with All Dataframes in Global Environment}
\usage{
receiver_performance_report(output_file = "Receiver_Performance_Report.html")
}
\arguments{
\item{output_file}{The file path where the report should be saved (e.g., "report.html").}
}
\value{
Generates an HTML report.
}
\description{
This function creates a semi-automated report using all dataframes in the global environment.
}
