% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{fathom_authenticate}
\alias{fathom_authenticate}
\title{Authenticate to Fathom API}
\usage{
fathom_authenticate(username, password, client_id)
}
\arguments{
\item{username}{Fathom account email}

\item{password}{Fathom password}

\item{client_id}{AWS Cognito client ID}
}
\value{
Access token string
}
\description{
Logs into the Fathom Cognito authentication system and returns an access token.
}
