% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{fathom_authenticate}
\alias{fathom_authenticate}
\title{Authenticate with Fathom Central}
\usage{
fathom_authenticate(username, password)
}
\arguments{
\item{username}{Your Fathom Central username}

\item{password}{Your Fathom Central password}
}
\value{
A session token (string)
}
\description{
This function logs in to Fathom Central and retrieves a session token.
}
