% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_detections.R
\name{get_detections}
\alias{get_detections}
\title{Get detections by species from Fathom Central}
\usage{
get_detections(species)
}
\arguments{
\item{species}{Character string of species name to filter}
}
\value{
Data frame with detections (placeholder)
}
\description{
Get detections by species from Fathom Central
}
