% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fathom_files.R
\name{parse_fathom_files}
\alias{parse_fathom_files}
\title{Parse Fathom Files}
\usage{
parse_fathom_files(
  directory_path,
  output = c("list", "env"),
  envir = .GlobalEnv,
  vars = "All",
  overwrite = TRUE
)
}
\arguments{
\item{output}{defines the output type "list" or "env"}

\item{envir}{defines the environment to which dataframes will be added}

\item{vars}{defines which Fathom variables to open. Options include: attitude, battery, cfg_channel, cfg_station,
 cfg_study, cfg_transmitter, clock_ref, data_source_file, depth, det, diag, event, event_init, event_offload, health_vr2ar, temp. All 
variables are loaded by default.}

\item{overwrite}{will overwrite existing objects in the environment with the same name}

\item{path}{path to the folder on your local machine containing the Fathom
interleaved .csv files}
}
\value{
Returns a list of dataframes, one for each record type contained within the
interleaved Fathom csv files, appending records from each receiver to the
appropriate dataframes. Output option "env" will return individual dataframes for each record type direct to the environment
}
\description{
The function parses Innovasea's Fathom csv schema log files and returns a named list
of dataframes, one for each record type. Fathom receiver log files (.vrl, .vdat) can be
converted to user readable formats using FathomR::convert_log_files() or
using Innovasea's Fathom Connect desktop application or cloud based service
Fathom Central. Note: Export formats from Fathom Central are restricted
interleaved .csv.
}
\details{
The *Date Time (UTC)* time variable is referenced to UTC time zone and is
not adjusted for clock drift. *Time* variable is equivalent to the *Date
Time (UTC)* following correction for clock drift using the *Time Correction
(s)* variable.
}
\examples{

\dontrun{
#'   # This example shows how to call the process_log_files function
#'   # Replace "/path/to/your/data" with a valid directory path on your system
#'   results <- parse_fathom_files('/path_to_fathom_csv_files')
#'   or
#'   parse_fathom_files('/path_to_fathom_csv_files', output ='env")
#' }


}
